refactor(transcriber): move prerecorded transcription into prerecorded/v2
Relocates the async (prerecorded) product into a versioned subpackage
mirroring sync/v1 and streaming/v3:

- transcriber.py -> prerecorded/v2/{transcript,transcript_group,client}.py,
  one public class per module beside its private impl
- transcript-only endpoints split out of api.py into prerecorded/v2/api.py;
  shared (upload_file, _get_error_message) and LeMUR pieces stay at root
- flat transcriber.py and api.py silently re-export the old surface, so
  every existing import path keeps resolving to identical objects
- adds test_transcriber_backwards_compat.py pinning old/new path identity

(v0.64.34)
