+++ assemblyai/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.64.33"
+__version__ = "0.64.34"
