+++ assemblyai/prerecorded/v2/__init__.py
@@ -1,0 +1,13 @@
+"""Prerecorded (async job) transcription against the v2 transcript API."""
+
+from ...types import TranscriptionConfig
+from .client import Transcriber
+from .transcript import Transcript
+from .transcript_group import TranscriptGroup
+
+__all__ = [
+    "Transcriber",
+    "Transcript",
+    "TranscriptGroup",
+    "TranscriptionConfig",
+]
